requirements.txt
python-telegram-bot>=21.0,<22.0
anthropic>=0.42.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
